# docs/openapi/novel-platform.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Novel Platform API",
    "version": "0.1.0",
    "description": "Versioned REST contract for the novel platform backend. Browser clients use the same-origin Next BFF: replace the backend prefix `/api/v1` with `/api/novel` for the `public`, `account`, `author`, and `admin` route groups. The BFF owns browser cookies, CSRF validation, and the upstream internal headers; do not call protected backend routes from browser JavaScript directly. `/api/v1/auth/*` is an internal BFF-to-backend contract and is intentionally not exposed by the catch-all browser proxy. This contract describes implemented local platform behavior only. Phone/email verification, WeChat/QQ OAuth, external payment, and production Qwen credentials are not represented as available API integrations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Backend service. Protected and auth routes require BFF-only headers."
    }
  ],
...
  "paths": {
    "/api/v1/public/books": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "searchPublishedBooks",
        "summary": "Search and filter published books",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/CatalogPage",
        "parameters": [
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/Category" },
          { "$ref": "#/components/parameters/CategoryAlias" },
          { "$ref": "#/components/parameters/SerialStatus" },
          { "$ref": "#/components/parameters/MinWords" },
          { "$ref": "#/components/parameters/MaxWords" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" } }
      }
    },
    "/api/v1/public/books/{id}": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getPublishedBook",
        "summary": "Get one published book with its published chapters and visible comments",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/BookDetail",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/api/v1/public/hot": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getHotBooks",
        "summary": "Get hot published books",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/BookList",
        "parameters": [{ "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/api/v1/public/recommendations": {
      "get": {
        "tags": ["Public discovery"],
        "operationId": "getRecommendedBooks",
        "summary": "Get editor-selected published books",
        "x-access": "public",
        "security": [],
        "x-response-data-schema": "#/components/schemas/BookList",
        "parameters": [{ "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/api/v1/author/reward-records": {
      "get": {
        "tags": ["Author workspace"],
        "operationId": "getAuthorRewardRecords",
        "summary": "Page successful token rewards for author-owned books",
        "description": "`from` and `to` are inclusive `Asia/Shanghai` calendar days. The summary covers every matching successful token reward, not only the requested page. Amounts are tokens only.",
        "security": [{ "BffSession": [] }],
        "x-required-role": "AUTHOR",
        "x-response-data-schema": "#/components/schemas/AuthorRewardReport",
        "parameters": [{ "$ref": "#/components/parameters/BookIdQuery" }, { "$ref": "#/components/parameters/FromDate" }, { "$ref": "#/components/parameters/ToDate" }, { "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/AuthorReportPageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
    "/api/v1/admin/editorial/recommendations": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getEditorialRecommendations",
        "summary": "Get persisted editorial recommendations",
        "security": [{ "BffSession": [] }],
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EditorialRecommendationList",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Operations"],
        "operationId": "assignEditorialRecommendation",
        "summary": "Assign a published book to an editorial rank",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EditorialRecommendation",
        "requestBody": { "$ref": "#/components/requestBodies/RecommendationAssignmentRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/admin/editorial/recommendations/{bookId}": {
      "put": {
        "tags": ["Operations"],
        "operationId": "reorderEditorialRecommendation",
        "summary": "Change an editorial recommendation rank",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EditorialRecommendation",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "requestBody": { "$ref": "#/components/requestBodies/RecommendationRankRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      },
      "delete": {
        "tags": ["Operations"],
        "operationId": "removeEditorialRecommendation",
        "summary": "Remove an editorial recommendation",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EmptyData",
        "parameters": [{ "$ref": "#/components/parameters/PathBookId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/admin/editorial/recommendations/audits": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getEditorialRecommendationAudits",
        "summary": "Get bounded editorial recommendation history",
        "security": [{ "BffSession": [] }],
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EditorialRecommendationAuditList",
        "parameters": [{ "$ref": "#/components/parameters/AuditLimit" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/admin/hot-searches": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getHotSearchTerms",
        "summary": "Get all managed hot-search terms",
        "security": [{ "BffSession": [] }],
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/HotSearchTermList",
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Operations"],
        "operationId": "createHotSearchTerm",
        "summary": "Create a managed hot-search term",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/HotSearchTerm",
        "requestBody": { "$ref": "#/components/requestBodies/HotSearchTermCreateRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/api/v1/admin/hot-searches/{termId}": {
      "put": {
        "tags": ["Operations"],
        "operationId": "updateHotSearchTerm",
        "summary": "Update a managed hot-search term",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/HotSearchTerm",
        "parameters": [{ "$ref": "#/components/parameters/PathTermId" }],
        "requestBody": { "$ref": "#/components/requestBodies/HotSearchTermUpdateRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" } }
      },
      "delete": {
        "tags": ["Operations"],
        "operationId": "removeHotSearchTerm",
        "summary": "Remove a managed hot-search term",
        "security": [{ "BffSession": [], "CsrfToken": [] }],
        "x-bff-csrf-required": true,
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/EmptyData",
        "parameters": [{ "$ref": "#/components/parameters/PathTermId" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/api/v1/admin/hot-searches/audits": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getHotSearchTermAudits",
        "summary": "Get bounded hot-search change history",
        "security": [{ "BffSession": [] }],
        "x-required-role": "ADMIN",
        "x-response-data-schema": "#/components/schemas/HotSearchTermAuditList",
        "parameters": [{ "$ref": "#/components/parameters/AuditLimit" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "novel_session",
        "description": "Opaque browser BFF session cookie. The browser never receives the backend session id."
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-CSRF",
        "description": "Required by the BFF on unsafe browser requests and must match its CSRF cookie/session binding."
      },
      "InternalBffKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Internal-Key",
        "description": "Private BFF-to-backend credential. Never send this header from a browser."
      },
      "BackendBffSession": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Novel-Bff-Session",
        "description": "Server-side backend session credential held by the BFF only. Never expose it to browser code."
      }
    },
    "parameters": {
      "PathId": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "PathBookId": { "name": "bookId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "PathTermId": { "name": "termId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "SearchQuery": { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Literal keyword search." },
      "Category": { "name": "category", "in": "query", "schema": { "type": "string" } },
      "CategoryAlias": { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "Compatibility alias used only when `category` is blank." },
      "SerialStatus": { "name": "status", "in": "query", "schema": { "type": "string" } },
      "MinWords": { "name": "minWords", "in": "query", "schema": { "type": "integer", "minimum": 0 } },
      "MaxWords": { "name": "maxWords", "in": "query", "schema": { "type": "integer", "minimum": 0 } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1 } },
...
      "BookIdQuery": { "name": "bookId", "in": "query", "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0, "maximum": 100000 } },
...
      "AuthorReportPageSize": { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
...
      "FromDate": { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
      "ToDate": { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
...
      "AuditLimit": { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
...
    },
    "requestBodies": {
...
      "RecommendationAssignmentRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecommendationAssignmentRequest" } } } },
      "RecommendationRankRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecommendationRankRequest" } } } },
      "HotSearchTermCreateRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HotSearchTermCreateRequest" } } } },
      "HotSearchTermUpdateRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HotSearchTermUpdateRequest" } } } }
    },
    "responses": {
      "Success": { "description": "Successful platform envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } } },
      "BadRequest": { "description": "Invalid request envelope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "Missing, expired, or invalid BFF session/internal BFF credential.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "Authenticated principal lacks the required role or ownership.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "Requested resource does not exist or is not visible to the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Requested state transition conflicts with current durable state.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer", "example": 200 }, "msg": { "type": "string", "example": "ok" }, "data": {} }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "msg", "data"],
        "properties": { "code": { "type": "integer" }, "msg": { "type": "string" }, "data": { "type": "null" } }
      },
      "Book": {
        "type": "object",
        "required": ["id", "title", "author", "category", "words", "serialStatus", "synopsis", "cover", "status", "authorId", "heat", "purchasePrice"],
        "properties": {
          "id": { "type": "integer", "format": "int64" }, "title": { "type": "string" }, "author": { "type": "string" }, "category": { "type": "string" }, "words": { "type": "integer" }, "serialStatus": { "type": "string" }, "synopsis": { "type": "string" }, "cover": { "type": "string" }, "status": { "type": "string", "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED", "NEEDS_REVIEW", "REJECTED", "OFFLINE"] }, "authorId": { "type": "integer", "format": "int64" }, "heat": { "type": "integer", "format": "int64" }, "purchasePrice": { "type": "integer", "format": "int64", "description": "Server-owned platform-token amount." }
        }
      },
      "BookList": { "type": "array", "items": { "$ref": "#/components/schemas/Book" } },
...
      "ChapterList": { "type": "array", "items": { "$ref": "#/components/schemas/Chapter" } },
      "Comment": { "type": "object", "required": ["id", "bookId", "userId", "authorName", "content", "status", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "chapterId": { "type": ["integer", "null"], "format": "int64" }, "userId": { "type": "integer", "format": "int64" }, "authorName": { "type": "string" }, "content": { "type": "string" }, "status": { "type": "string", "enum": ["PENDING_REVIEW", "VISIBLE", "REJECTED"] }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "HotSearchTerm": { "type": "object", "required": ["id", "term", "enabled", "rank", "createdAt", "updatedAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "term": { "type": "string" }, "enabled": { "type": "boolean" }, "rank": { "type": "integer" }, "createdByUserId": { "type": ["integer", "null"], "format": "int64" }, "updatedByUserId": { "type": ["integer", "null"], "format": "int64" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "HotSearchTermList": { "type": "array", "items": { "$ref": "#/components/schemas/HotSearchTerm" } },
      "DiscoveryFacets": { "type": "object", "properties": { "categories": { "type": "array", "items": { "type": "string" } }, "serialStatuses": { "type": "array", "items": { "type": "string" } }, "wordCountRanges": { "type": "array", "items": { "$ref": "#/components/schemas/WordCountRange" } } } },
      "WordCountRange": { "type": "object", "required": ["key", "label"], "properties": { "key": { "type": "string" }, "label": { "type": "string" }, "minWords": { "type": ["integer", "null"] }, "maxWords": { "type": ["integer", "null"] } } },
      "CatalogPage": { "type": "object", "required": ["items", "meta"], "properties": { "items": { "$ref": "#/components/schemas/BookList" }, "meta": { "type": "object", "required": ["total", "facets", "query"], "properties": { "total": { "type": "integer", "format": "int64" }, "facets": { "$ref": "#/components/schemas/DiscoveryFacets" }, "query": { "type": "object" } } } } },
      "BookDetail": { "type": "object", "required": ["book", "chapters", "comments"], "properties": { "book": { "$ref": "#/components/schemas/Book" }, "chapters": { "$ref": "#/components/schemas/ChapterList" }, "comments": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } } } },
...
      "AuthorRewardReport": { "type": "object", "required": ["items", "summary", "meta"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AuthorRewardRecord" } }, "summary": { "$ref": "#/components/schemas/AuthorRewardSummary" }, "meta": { "$ref": "#/components/schemas/AuthorRewardMetadata" } } },
      "AuthorRewardRecord": { "type": "object", "required": ["id", "bookId", "bookTitle", "rewarderUserId", "tokenAmount", "rewardedAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "bookTitle": { "type": "string" }, "rewarderUserId": { "type": "integer", "format": "int64" }, "tokenAmount": { "type": "integer", "format": "int64" }, "rewardedAt": { "type": "string", "format": "date-time" } } },
      "AuthorRewardSummary": { "type": "object", "required": ["rewardCount", "totalTokens", "amountUnit"], "properties": { "rewardCount": { "type": "integer", "format": "int64" }, "totalTokens": { "type": "integer", "format": "int64" }, "amountUnit": { "type": "string", "const": "TOKEN" } } },
      "AuthorRewardMetadata": { "type": "object", "required": ["total", "page", "size", "timeZone", "dateBoundary", "recordInclusion"], "properties": { "total": { "type": "integer", "format": "int64" }, "page": { "type": "integer" }, "size": { "type": "integer" }, "bookId": { "type": ["integer", "null"], "format": "int64" }, "from": { "type": ["string", "null"], "format": "date" }, "to": { "type": ["string", "null"], "format": "date" }, "timeZone": { "type": "string", "const": "Asia/Shanghai" }, "dateBoundary": { "type": "string", "const": "FROM_INCLUSIVE_TO_INCLUSIVE" }, "recordInclusion": { "type": "string", "const": "SUCCESSFUL_BOOK_REWARD_DEBIT_ONLY" } } },
...
      "EditorialRecommendation": { "type": "object", "required": ["book", "rank"], "properties": { "book": { "$ref": "#/components/schemas/Book" }, "rank": { "type": "integer" } } },
      "EditorialRecommendationList": { "type": "array", "items": { "$ref": "#/components/schemas/EditorialRecommendation" } },
      "EditorialRecommendationAudit": { "type": "object", "required": ["id", "bookId", "action", "details", "operatorUserId", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "bookId": { "type": "integer", "format": "int64" }, "action": { "type": "string" }, "previousRank": { "type": ["integer", "null"] }, "rank": { "type": ["integer", "null"] }, "details": { "type": "string" }, "operatorUserId": { "type": "integer", "format": "int64" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "EditorialRecommendationAuditList": { "type": "array", "items": { "$ref": "#/components/schemas/EditorialRecommendationAudit" } },
      "HotSearchTermAudit": { "type": "object", "required": ["id", "termId", "term", "action", "details", "operatorUserId", "createdAt"], "properties": { "id": { "type": "integer", "format": "int64" }, "termId": { "type": "integer", "format": "int64" }, "term": { "type": "string" }, "action": { "type": "string" }, "previousRank": { "type": ["integer", "null"] }, "rank": { "type": ["integer", "null"] }, "details": { "type": "string" }, "operatorUserId": { "type": "integer", "format": "int64" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "HotSearchTermAuditList": { "type": "array", "items": { "$ref": "#/components/schemas/HotSearchTermAudit" } },
...
      "EmptyData": { "type": "null" },
...
      "RecommendationAssignmentRequest": { "type": "object", "required": ["bookId"], "properties": { "bookId": { "type": "integer", "format": "int64", "minimum": 1 }, "rank": { "type": "integer", "minimum": 1, "maximum": 100000 } }, "additionalProperties": false },
      "RecommendationRankRequest": { "type": "object", "required": ["rank"], "properties": { "rank": { "type": "integer", "minimum": 1, "maximum": 100000 } }, "additionalProperties": false },
      "HotSearchTermCreateRequest": { "type": "object", "required": ["term", "enabled"], "properties": { "term": { "type": "string", "minLength": 1, "maxLength": 100 }, "enabled": { "type": "boolean" }, "rank": { "type": "integer", "minimum": 1, "maximum": 100000 } }, "additionalProperties": false },
      "HotSearchTermUpdateRequest": { "type": "object", "required": ["term", "enabled", "rank"], "properties": { "term": { "type": "string", "minLength": 1, "maxLength": 100 }, "enabled": { "type": "boolean" }, "rank": { "type": "integer", "minimum": 1, "maximum": 100000 } }, "additionalProperties": false }
    }
  }
}